\name{dot-PBSddeEnv}
\docType{data}
\alias{.PBSddeEnv}
\title{
  PBSddesolve Environment
}
\description{
  An environment set aside for \code{PBSddesolve}.
}
\usage{.PBSddeEnv}
\format{
A new environment with a \code{.GlobalEnv} parent.
}
\source{
  Generated by a call to the base function \code{new.env()}.
}
\details{
  The environment is created in \code{'zzz.r'} and can be used by
  \code{PBSmodelling} functions \code{'lisp'}, \code{'tget'}, 
  \code{'tput'}, \code{'tprint'}, and \code{'tcall'}.
}
\seealso{
In \pkg{PBSmodelling}:\cr
  \code{\link[PBSmodelling]{lisp}},
  \code{\link[PBSmodelling]{tget}}
}
\keyword{environment}

